refactor server models to mod
