+++ src/server/models/mod.rs
@@ -1,0 +1,2 @@
+pub mod dispatcher;
+pub mod room;
